Fix GCC -Wconversion warnings

+++ pde/src/sgpp/pde/application/HeatEquationSolver.cpp
@@ -36,7 +36,7 @@ HeatEquationSolver::~HeatEquationSolver() {
 
 void HeatEquationSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<size_t>(level);
 
   this->myGrid = new base::LinearBoundaryGrid(BoundingBox);
